// migrations/snapshots/cfaab9ad41cee3b1aa70ed2d7d812d7318f9f799805742a0ae451dd49bb4f3aa/contract.d.ts
export type CodecTypes = PgTypes;
export type LaneCodecTypes = CodecTypes;
export type QueryOperationTypes = PgAdapterQueryOps<CodecTypes>;
// ...
export type FieldOutputTypes = {
  readonly public: {
    readonly Category: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
      readonly slug: CodecTypes['pg/text@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamp-temporal@1']['output'];
    };
    readonly Horoscope: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly zodiacSign:
        | 'ARIES'
        | 'TAURUS'
        | 'GEMINI'
        | 'CANCER'
        | 'LEO'
        | 'VIRGO'
        | 'LIBRA'
        | 'SCORPIO'
        | 'SAGITTARIUS'
        | 'CAPRICORN'
        | 'AQUARIUS'
        | 'PISCES';
      readonly date: CodecTypes['pg/timestamp-temporal@1']['output'];
      readonly description: CodecTypes['pg/text@1']['output'];
      readonly luckyColor: CodecTypes['pg/text@1']['output'] | null;
      readonly luckyNumber: CodecTypes['pg/int4@1']['output'] | null;
      readonly mood: CodecTypes['pg/text@1']['output'] | null;
      readonly createdAt: CodecTypes['pg/timestamp-temporal@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamp-temporal@1']['output'];
    };
    readonly PastEvent: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly month: CodecTypes['pg/int4@1']['output'];
      readonly day: CodecTypes['pg/int4@1']['output'];
      readonly year: CodecTypes['pg/int4@1']['output'];
      readonly title: CodecTypes['pg/text@1']['output'];
      readonly slug: CodecTypes['pg/text@1']['output'];
      readonly description: CodecTypes['pg/text@1']['output'];
      readonly category:
        | 'war'
        | 'discovery'
        | 'politics'
        | 'science'
        | 'art'
        | 'sports'
        | 'technology'
        | 'medicine'
        | 'exploration'
        | 'literature'
        | 'music'
        | 'economy'
        | 'religion'
        | 'disaster'
        | 'revolution'
        | 'invention';
      readonly tags: ReadonlyArray<CodecTypes['pg/text@1']['output']>;
      readonly sourceUrls: ReadonlyArray<CodecTypes['pg/text@1']['output']>;
      readonly metadata: CodecTypes['pg/jsonb@1']['output'] | null;
      readonly isPublished: CodecTypes['pg/bool@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamp-temporal@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamp-temporal@1']['output'];
      readonly eventDate: CodecTypes['pg/timestamp-temporal@1']['output'] | null;
    };
    readonly Question: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly quizId: CodecTypes['pg/text@1']['output'];
      readonly text: CodecTypes['pg/text@1']['output'];
      readonly options: ReadonlyArray<CodecTypes['pg/text@1']['output']>;
      readonly correctIndex: CodecTypes['pg/int4@1']['output'];
      readonly explanation: CodecTypes['pg/text@1']['output'] | null;
      readonly createdAt: CodecTypes['pg/timestamp-temporal@1']['output'];
    };
    readonly Quiz: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly quizPageTitle: CodecTypes['pg/text@1']['output'];
      readonly quizPageDescription: CodecTypes['pg/text@1']['output'];
      readonly difficulty: 'easy' | 'medium' | 'hard';
      readonly title: CodecTypes['pg/text@1']['output'];
      readonly description: CodecTypes['pg/text@1']['output'];
      readonly slug: CodecTypes['pg/text@1']['output'];
      readonly isPublished: CodecTypes['pg/bool@1']['output'];
      readonly publishedAt: CodecTypes['pg/timestamp-temporal@1']['output'] | null;
      readonly views: CodecTypes['pg/int4@1']['output'];
      readonly categoryId: CodecTypes['pg/int4@1']['output'] | null;
      readonly createdAt: CodecTypes['pg/timestamp-temporal@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamp-temporal@1']['output'];
      readonly subCategoryId: CodecTypes['pg/int4@1']['output'] | null;
    };
    readonly QuizTag: {
      readonly quizId: CodecTypes['pg/text@1']['output'];
      readonly tagId: CodecTypes['pg/int4@1']['output'];
    };
    readonly SubCategory: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
      readonly slug: CodecTypes['pg/text@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamp-temporal@1']['output'];
      readonly categoryId: CodecTypes['pg/int4@1']['output'];
    };
    readonly Tag: {
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamp-temporal@1']['output'];
    };
  };
};
export type FieldInputTypes = {
  readonly public: {
    readonly Category: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
      readonly slug: CodecTypes['pg/text@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamp-temporal@1']['input'];
    };
    readonly Horoscope: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly zodiacSign:
        | 'ARIES'
        | 'TAURUS'
        | 'GEMINI'
        | 'CANCER'
        | 'LEO'
        | 'VIRGO'
        | 'LIBRA'
        | 'SCORPIO'
        | 'SAGITTARIUS'
        | 'CAPRICORN'
        | 'AQUARIUS'
        | 'PISCES';
      readonly date: CodecTypes['pg/timestamp-temporal@1']['input'];
      readonly description: CodecTypes['pg/text@1']['input'];
      readonly luckyColor: CodecTypes['pg/text@1']['input'] | null;
      readonly luckyNumber: CodecTypes['pg/int4@1']['input'] | null;
      readonly mood: CodecTypes['pg/text@1']['input'] | null;
      readonly createdAt: CodecTypes['pg/timestamp-temporal@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamp-temporal@1']['input'];
    };
    readonly PastEvent: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly month: CodecTypes['pg/int4@1']['input'];
      readonly day: CodecTypes['pg/int4@1']['input'];
      readonly year: CodecTypes['pg/int4@1']['input'];
      readonly title: CodecTypes['pg/text@1']['input'];
      readonly slug: CodecTypes['pg/text@1']['input'];
      readonly description: CodecTypes['pg/text@1']['input'];
      readonly category:
        | 'war'
        | 'discovery'
        | 'politics'
        | 'science'
        | 'art'
        | 'sports'
        | 'technology'
        | 'medicine'
        | 'exploration'
        | 'literature'
        | 'music'
        | 'economy'
        | 'religion'
        | 'disaster'
        | 'revolution'
        | 'invention';
      readonly tags: ReadonlyArray<CodecTypes['pg/text@1']['input']>;
      readonly sourceUrls: ReadonlyArray<CodecTypes['pg/text@1']['input']>;
      readonly metadata: CodecTypes['pg/jsonb@1']['input'] | null;
      readonly isPublished: CodecTypes['pg/bool@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamp-temporal@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamp-temporal@1']['input'];
      readonly eventDate: CodecTypes['pg/timestamp-temporal@1']['input'] | null;
    };
    readonly Question: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly quizId: CodecTypes['pg/text@1']['input'];
      readonly text: CodecTypes['pg/text@1']['input'];
      readonly options: ReadonlyArray<CodecTypes['pg/text@1']['input']>;
      readonly correctIndex: CodecTypes['pg/int4@1']['input'];
      readonly explanation: CodecTypes['pg/text@1']['input'] | null;
      readonly createdAt: CodecTypes['pg/timestamp-temporal@1']['input'];
    };
    readonly Quiz: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly quizPageTitle: CodecTypes['pg/text@1']['input'];
      readonly quizPageDescription: CodecTypes['pg/text@1']['input'];
      readonly difficulty: 'easy' | 'medium' | 'hard';
      readonly title: CodecTypes['pg/text@1']['input'];
      readonly description: CodecTypes['pg/text@1']['input'];
      readonly slug: CodecTypes['pg/text@1']['input'];
      readonly isPublished: CodecTypes['pg/bool@1']['input'];
      readonly publishedAt: CodecTypes['pg/timestamp-temporal@1']['input'] | null;
      readonly views: CodecTypes['pg/int4@1']['input'];
      readonly categoryId: CodecTypes['pg/int4@1']['input'] | null;
      readonly createdAt: CodecTypes['pg/timestamp-temporal@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamp-temporal@1']['input'];
      readonly subCategoryId: CodecTypes['pg/int4@1']['input'] | null;
    };
    readonly QuizTag: {
      readonly quizId: CodecTypes['pg/text@1']['input'];
      readonly tagId: CodecTypes['pg/int4@1']['input'];
    };
    readonly SubCategory: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
      readonly slug: CodecTypes['pg/text@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamp-temporal@1']['input'];
      readonly categoryId: CodecTypes['pg/int4@1']['input'];
    };
    readonly Tag: {
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamp-temporal@1']['input'];
    };
  };
};
// ...
export type TypeMaps = TypeMapsType<
  CodecTypes,
  QueryOperationTypes,
  FieldOutputTypes,
  FieldInputTypes,
  StorageColumnTypes,
  StorageColumnInputTypes,
  AggregateTypes
>;
// ...
export type Contract = ContractWithTypeMaps<ContractBase, TypeMaps>;